--- a/fractal.js
+++ b/fractal.js
@@ -73,7 +73,7 @@ module.exports = {
     var vfOpenBrowser = typeof global.vfOpenBrowser === "undefined" ? true : global.vfOpenBrowser;
     fractal.web.set('server.syncOptions', {
       watchOptions: {
-        ignored: path.join(__dirname, '**/*.scss'),
+        ignored: path.join(__dirname, './components/**/*.scss'),
       },
       open: vfOpenBrowser,
       browser: 'default',