@khawkins98 2019-04-16
Also a speed bump on the MacBook Pro (from 19s to 14s).