@sturobson 2019-04-16
mind if I add 

```
    watchOptions: {
        ignored: path.join(__dirname, '**/*.scss'),
      },
```

to `fractal.web.set('server.syncOptions', {}` in `fractal.js` so the page doesn't reload on Sass file changes - but the Sass changes get injected still?