@khawkins98 2019-04-16
w00t, down to 5s on the MB Pro.